[EC-507] Move DeleteGroupCommand to OrganizationFeatures/OrganizationUsers
