+++ bitwarden_license/src/Scim/Controllers/v2/GroupsController.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.Json;
 using Bit.Core.Entities;
+using Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
-using Bit.Scim.Commands.Groups.Interfaces;
 using Bit.Scim.Context;
 using Bit.Scim.Models;
 using Bit.Scim.Utilities;
