+++ bitwarden_license/src/Scim/Users/GetUserQuery.cs
@@ -1,9 +1,9 @@
 ﻿using Bit.Core.Exceptions;
 using Bit.Core.Repositories;
 using Bit.Scim.Models;
-using Bit.Scim.Queries.Users.Interfaces;
+using Bit.Scim.Users.Interfaces;
 
-namespace Bit.Scim.Queries.Users;
+namespace Bit.Scim.Users;
 
 public class GetUserQuery : IGetUserQuery
 {
