+++ bitwarden_license/test/Scim.Test/Users/GetUserQueryTests.cs
@@ -2,7 +2,7 @@
 using Bit.Core.Exceptions;
 using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 using Bit.Core.Repositories;
-using Bit.Scim.Queries.Users;
+using Bit.Scim.Users;
 using Bit.Scim.Utilities;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
@@ -10,7 +10,7 @@ using Bit.Test.Common.Helpers;
 using NSubstitute;
 using Xunit;
 
-namespace Bit.Scim.Test.Queries.Users;
+namespace Bit.Scim.Test.Users;
 
 [SutProviderCustomize]
 public class GetUserQueryTests
