+++ src/Core/Services/Implementations/GroupService.cs
@@ -75,6 +75,7 @@ public class GroupService : IGroupService
         }
     }
 
+    [Obsolete("IDeleteGroupCommand should be used instead. To be removed by EC-608.")]
     public async Task DeleteAsync(Group group)
     {
         await _groupRepository.DeleteAsync(group);
